EXTENSIVE WORK TO ABOUT ME SECTION
Added a timer that counts up from the day that I started programming.

Seriously fixed up the height calculation so that, preferably, the
"about me" section will be exactly the height of the browser.

Added a hook within sections.js for specific sections, so that it's
easy to create custom code for each one, and override functionality.
(In this case, I had to override the default height functionality for
the about me section.)